Shadcn-style badges, move filters to header, labels to title row
- Restyle filters and post labels as rounded pill badges (shadcn outline style)
- Move filter controls into the collapsing header subheader (currently hidden)
- Move post tag labels inline with title, right-aligned
- Lift filter state to App.js, add showTitle prop to Post
- Add italic "prev." prefix to prior job roles in header
- Flex layout for header subheader: bio left, filters right
