+++ src/GridFeed.css
@@ -10,31 +10,35 @@
 }
 
 .grid-feed-filters {
-  display: flex;
-  gap: 0.5rem;
+  display: none;
+  gap: 0.4rem;
   flex-wrap: wrap;
-  margin: 1.5rem 0 1rem;
+  justify-content: flex-end;
+  margin: 0.75rem 0 0.25rem;
 }
 
 .grid-feed-tag {
   font-family: var(--font-body);
-  font-size: 0.85rem;
+  font-size: 0.75rem;
   background: transparent;
   color: var(--color-text);
-  border: 1px solid var(--color-muted);
-  padding: 0.3rem 0.75rem;
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  padding: 0.2rem 0.6rem;
+  border-radius: 9999px;
   cursor: pointer;
-  transition: background-color 0.2s, border-color 0.2s;
+  transition: background-color 0.2s, border-color 0.2s, color 0.2s;
+  line-height: 1.4;
 }
 
 .grid-feed-tag:hover {
-  border-color: var(--color-accent);
+  border-color: rgba(255, 255, 255, 0.5);
+  background: rgba(255, 255, 255, 0.05);
 }
 
 .grid-feed-tag.active {
   background-color: var(--color-accent);
   border-color: var(--color-accent);
-  color: var(--color-bg);
+  color: #fff;
 }
 
 .timeline {
@@ -68,12 +72,6 @@
   background: var(--timeline-line-color);
 }
 
-.timeline-entry:last-child .timeline-rail::before {
-  bottom: auto;
-  height: calc(var(--timeline-sticky-top) - 2.75rem);
-  min-height: 2rem;
-}
-
 .timeline-marker {
   position: sticky;
   top: var(--timeline-sticky-top);
@@ -81,7 +79,7 @@
   display: inline-flex;
   align-items: center;
   gap: 0.75rem;
-  padding: 0.75rem 0 var(--timeline-line-gap);
+  padding: 0.5rem 0;
   background: var(--color-bg);
 }
 
@@ -107,22 +105,37 @@
   line-height: 1;
 }
 
+.timeline-title-row {
+  display: flex;
+  align-items: baseline;
+  justify-content: space-between;
+  gap: 0.75rem;
+  margin-top: 0;
+}
+
+.timeline-title-row h3 {
+  margin-top: 0;
+  margin-bottom: 1rem;
+  font-size: inherit;
+  white-space: nowrap;
+}
+
 .timeline-post-tags {
   display: flex;
-  gap: 0.4rem;
+  gap: 0.25rem;
   flex-wrap: wrap;
-  min-height: 1.35rem;
-  margin-bottom: 0.5rem;
 }
 
 .timeline-tag-badge {
-  font-size: 0.7rem;
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-  padding: 0.15rem 0.5rem;
-  border: 1px solid var(--color-muted);
+  font-family: var(--font-body);
+  font-size: 0.75rem;
+  background: transparent;
   color: var(--color-text);
-  border-radius: 2px;
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  padding: 0.2rem 0.6rem;
+  border-radius: 9999px;
+  line-height: 1.4;
+  white-space: nowrap;
 }
 
 .timeline-content .post {
@@ -146,5 +159,10 @@
 
   .timeline-date {
     font-size: 0.75rem;
+  }
+
+  .timeline-tag-badge {
+    font-size: 0.65rem;
+    padding: 0.15rem 0.45rem;
   }
 }
